No need for dedicated resource list RPC message definition
