+++ code/framework/src/integrations/server/instance.cpp
@@ -18,7 +18,6 @@
 #include "networking/messages/client_ready_assets.h"
 #include "networking/messages/client_request_streamer.h"
 #include "networking/messages/messages.h"
-#include "networking/messages/resource_list.h"
 #include "integrations/shared/rpc/emit_lua_event.h"
 
 #include "scripting/builtins/entity.h"
@@ -292,12 +291,14 @@ namespace Framework::Integrations::Server {
             }
 
             // Let the client know they can ask for client-side assets now.
+            // Include the resource list in the message.
             ClientReadyAssets readyMsg;
-            net->Send(readyMsg, guid);
-
             if (_scriptingModule) {
-                _scriptingModule->SendResourceListToClient(guid);
+                for (const auto &resource : _scriptingModule->GetClientResourceList()) {
+                    readyMsg.AddResource(resource.name, resource.version, resource.hash);
+                }
             }
+            net->Send(readyMsg, guid);
         });
 
         net->SetOnPlayerDisconnectCallback([this, net](SLNet::Packet *packet, uint32_t reason) {
